move setup functions for grid maker and example provider to new file
