+++ gnina/setup.py
@@ -1,0 +1,63 @@
+"""
+Utility functions to setup training and inference.
+"""
+
+import molgrid
+
+# TODO: Ensure that for inference all systems are seen exactly once?
+_iteration_schemes = {
+    "small": molgrid.IterationScheme.SmallEpoch,
+    "large": molgrid.IterationScheme.LargeEpoch,
+}
+
+
+def setup_example_provider(examples_file, args) -> molgrid.ExampleProvider:
+    """
+    Setup :code:`molgrid.ExampleProvider` based on command line arguments.
+
+    Parameters
+    ----------
+    examples_file: str
+        File with examples (.types file)
+    args:
+        Command line arguments
+
+    Returns
+    -------
+    molgrid.ExampleProvider
+        Initialized :code:`molgrid.ExampleProvider`
+    """
+    example_provider = molgrid.ExampleProvider(
+        data_root=args.data_root,
+        balanced=args.balanced,
+        shuffle=args.shuffle,
+        default_batch_size=args.batch_size,
+        iteration_scheme=_iteration_schemes[args.iteration_scheme],
+        ligmolcache=args.ligmolcache,
+        recmolcache=args.recmolcache,
+        stratify_receptor=args.stratify_receptor,
+        cache_structs=True,
+    )
+    example_provider.populate(examples_file)
+
+    return example_provider
+
+
+def setup_grid_maker(args) -> molgrid.GridMaker:
+    """
+    Setup :code:`molgrid.ExampleProvider` and :code:`molgrid.GridMaker` based on command
+    line arguments.
+
+    Parameters
+    ----------
+    args:
+        Command line arguments
+
+    Returns
+    -------
+    molgrid.GridMaker
+        Initialized :code:`molgrid.GridMaker`
+    """
+    grid_maker = molgrid.GridMaker(resolution=args.resolution, dimension=args.dimension)
+
+    return grid_maker
